requirements.txt
discord.py
fpdf
requests
flask
